Move manually tracked balances code into its own module
